Move setops tests to test_setops.py and add two tests
